// api/client/node/src/schemas.d.ts
export interface paths {
// ...
  '/api/v1/plans': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * List plans
     * @description List all plans.
     */
    get: operations['listPlans']
    put?: never
    /**
     * Create a plan
     * @description Create a new plan.
     */
    post: operations['createPlan']
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
  '/api/v1/plans/{planId}': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    /**
     * Get plan
     * @description Get a plan by id or key. The latest published version is returned if latter is used.
     */
    get: operations['getPlan']
    /**
     * Update a plan
     * @description Update plan by id.
     */
    put: operations['updatePlan']
    post?: never
    /**
     * Delete plan
     * @description Soft delete plan by plan.id.
     *
     *     Once a plan is deleted it cannot be undeleted.
     */
    delete: operations['deletePlan']
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  '/api/v1/plans/{planId}/archive': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    get?: never
    put?: never
    /**
     * Archive plan version
     * @description Archive a plan version.
     */
    post: operations['archivePlan']
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
  '/api/v1/plans/{planId}/publish': {
    parameters: {
      query?: never
      header?: never
      path?: never
      cookie?: never
    }
    get?: never
    put?: never
    /**
     * Publish plan
     * @description Publish a plan version.
     */
    post: operations['publishPlan']
    delete?: never
    options?: never
    head?: never
    patch?: never
    trace?: never
  }
// ...
}
// ...
export type BadRequestProblemResponse =
  components['schemas']['BadRequestProblemResponse']
// ...
export type ForbiddenProblemResponse =
  components['schemas']['ForbiddenProblemResponse']
// ...
export type InternalServerErrorProblemResponse =
  components['schemas']['InternalServerErrorProblemResponse']
// ...
export type NotFoundProblemResponse =
  components['schemas']['NotFoundProblemResponse']
// ...
export type Plan = components['schemas']['Plan']
export type PlanCreate = components['schemas']['PlanCreate']
export type PlanOrderBy = components['schemas']['PlanOrderBy']
export type PlanPaginatedResponse =
  components['schemas']['PlanPaginatedResponse']
export type PlanPhase = components['schemas']['PlanPhase']
export type PlanReference = components['schemas']['PlanReference']
export type PlanReferenceInput = components['schemas']['PlanReferenceInput']
export type PlanReplaceUpdate = components['schemas']['PlanReplaceUpdate']
// ...
export type UnauthorizedProblemResponse =
  components['schemas']['UnauthorizedProblemResponse']
// ...
export interface operations {
// ...
  listPlans: {
    parameters: {
      query?: {
        /** @description Include deleted plans in response.
         *
         *     Usage: `?includeDeleted=true` */
        includeDeleted?: boolean
        /** @description Filter by plan.id attribute */
        id?: string[]
        /** @description Filter by plan.key attribute */
        key?: string[]
        /** @description Filter by plan.key and plan.version attributes */
        keyVersion?: {
          [key: string]: number[]
        }
        /** @description Start date-time in RFC 3339 format.
         *
         *     Inclusive. */
        page?: components['parameters']['Pagination.page']
        /** @description Number of items per page.
         *
         *     Default is 100. */
        pageSize?: components['parameters']['Pagination.pageSize']
        /** @description The order direction. */
        order?: components['parameters']['PlanOrderByOrdering.order']
        /** @description The order by field. */
        orderBy?: components['parameters']['PlanOrderByOrdering.orderBy']
      }
      header?: never
      path?: never
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description The request has succeeded. */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/json': components['schemas']['PlanPaginatedResponse']
        }
      }
      /** @description The server cannot or will not process the request due to something that is perceived to be a client error (e.g., malformed request syntax, invalid request message framing, or deceptive request routing). */
      400: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['BadRequestProblemResponse']
        }
      }
      /** @description The request has not been applied because it lacks valid authentication credentials for the target resource. */
      401: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['UnauthorizedProblemResponse']
        }
      }
      /** @description The server understood the request but refuses to authorize it. */
      403: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['ForbiddenProblemResponse']
        }
      }
      /** @description The server encountered an unexpected condition that prevented it from fulfilling the request. */
      500: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['InternalServerErrorProblemResponse']
        }
      }
      /** @description The server is currently unable to handle the request due to a temporary overload or scheduled maintenance, which will likely be alleviated after some delay. */
      503: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['ServiceUnavailableProblemResponse']
        }
      }
      /** @description An unexpected error response. */
      default: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['UnexpectedProblemResponse']
        }
      }
    }
  }
// ...
  getPlan: {
    parameters: {
      query?: {
        /** @description Include latest version of the Plan instead of the version in active state.
         *
         *     Usage: `?includeLatest=true` */
        includeLatest?: boolean
      }
      header?: never
      path: {
        planId: string
      }
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description The request has succeeded. */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/json': components['schemas']['Plan']
        }
      }
      /** @description The server cannot or will not process the request due to something that is perceived to be a client error (e.g., malformed request syntax, invalid request message framing, or deceptive request routing). */
      400: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['BadRequestProblemResponse']
        }
      }
      /** @description The request has not been applied because it lacks valid authentication credentials for the target resource. */
      401: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['UnauthorizedProblemResponse']
        }
      }
      /** @description The server understood the request but refuses to authorize it. */
      403: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['ForbiddenProblemResponse']
        }
      }
      /** @description The origin server did not find a current representation for the target resource or is not willing to disclose that one exists. */
      404: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['NotFoundProblemResponse']
        }
      }
      /** @description The server encountered an unexpected condition that prevented it from fulfilling the request. */
      500: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['InternalServerErrorProblemResponse']
        }
      }
      /** @description The server is currently unable to handle the request due to a temporary overload or scheduled maintenance, which will likely be alleviated after some delay. */
      503: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['ServiceUnavailableProblemResponse']
        }
      }
      /** @description An unexpected error response. */
      default: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['UnexpectedProblemResponse']
        }
      }
    }
  }
  updatePlan: {
    parameters: {
      query?: never
      header?: never
      path: {
        planId: string
      }
      cookie?: never
    }
    requestBody: {
      content: {
        'application/json': components['schemas']['PlanReplaceUpdate']
      }
    }
    responses: {
      /** @description The request has succeeded. */
      200: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/json': components['schemas']['Plan']
        }
      }
      /** @description The server cannot or will not process the request due to something that is perceived to be a client error (e.g., malformed request syntax, invalid request message framing, or deceptive request routing). */
      400: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['BadRequestProblemResponse']
        }
      }
      /** @description The request has not been applied because it lacks valid authentication credentials for the target resource. */
      401: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['UnauthorizedProblemResponse']
        }
      }
      /** @description The server understood the request but refuses to authorize it. */
      403: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['ForbiddenProblemResponse']
        }
      }
      /** @description The origin server did not find a current representation for the target resource or is not willing to disclose that one exists. */
      404: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['NotFoundProblemResponse']
        }
      }
      /** @description The server encountered an unexpected condition that prevented it from fulfilling the request. */
      500: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['InternalServerErrorProblemResponse']
        }
      }
      /** @description The server is currently unable to handle the request due to a temporary overload or scheduled maintenance, which will likely be alleviated after some delay. */
      503: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['ServiceUnavailableProblemResponse']
        }
      }
      /** @description An unexpected error response. */
      default: {
        headers: {
          [name: string]: unknown
        }
        content: {
          'application/problem+json': components['schemas']['UnexpectedProblemResponse']
        }
      }
    }
  }
// ...
}